Move the book card and grid out of the landing page
Both are about to be shared with a library page, so they no longer belong
to the page that happens to render them today. `RecentBooks` stays with
the landing page: it is a dashboard row, not a way to browse the library.
