+++ frontend/src/components/books/BookCard.tsx
