+++ frontend/src/pages/LandingPage/components/RecentBooks.tsx
@@ -1,11 +1,11 @@
 import { useGetRecentBooks } from '@/api/generated/books/books';
 import { Spinner } from '@/components/animations/Spinner.tsx';
+import { BookCard } from '@/components/books/BookCard.tsx';
 import { Carousel } from '@/components/carousel/Carousel.tsx';
 import { CarouselItem } from '@/components/carousel/CarouselItem.tsx';
 import { PAGE_GUTTER } from '@/components/layout/Layouts.tsx';
 import { SectionTitle } from '@/components/typography/SectionTitle.tsx';
 import { Alert, Box } from '@mui/material';
-import { BookCard } from './BookCard';
 
 const RECENT_BOOKS_LIMIT = 8;
